rerddapXtracto: Routines to simplify data extraction using ERD's ERDDAP web service.

The rerddapXtracto package subsets and extracts satellite and other oceanographic related data from any ERDDAP server using the R package "rerddap"developed by Scott Chamberlain and the wonderful people at "rOpenSci".

Details

The package contains three main functions:

Main Functions

rxtracto - Extracts an environmental variable along a track defined by its longtiude, latitude and time.

rxtracto_3D - Extracts an environmental variable in a 3D (longitude,latitude, time) bounding box

rxtractogon - Extracts an environmental variable in a spatial polygon through time.

These functions require a call be made to rerddap:info() for the appropriate ERDDAP server and datasetID. #' @section Details: Besides the terse help documents, more detail in using the functions are given in the included vignette "UsingrerddapXtracto". The datasets used in the vignette are included in the "data" directory.

See also

Author

Maintainer: Roy Mendelssohn roy.mendelssohn@noaa.gov

Other contributors:

Marie Auger-Methe [contributor]