plot result of 'rxtracto_3D'

plotBox is a function to plot the results from 'rxtracto()' and 'xtracto()'

plotBBox(resp, plotColor = "viridis", time = NA, myFunc = NA,
  mapData = NULL, crs = NULL, animate = FALSE, cumulative = FALSE,
  name = NA, maxpixels = 10000)

Arguments

resp
data frame returned from 'rxtracto()'

plotColor
the color to use in plot from 'rerddap'

time
a function to map multi-time to one, or else identity for animation

myFunc
function of one argument to transform the data

mapData
map data from 'maps' or 'mapdata', must be of class 'map'

crs
valid crs string

animate
if multiple times, if TRUE will animate the maps

cumulative
makes cumulative animation of data

name
name for colorbar label

maxpixels
maximum number of pixels to use in making the map - controls resolution

Value

a 'plotdap' plot

Examples

dataInfo <- rerddap::info('erdMBsstd1day')
parameter <- 'sst'
xcoord <- c(230, 232)
ycoord <- c(33, 35)
tcoord <- c('2006-01-15', '2006-01-15')
zcoord <- c(0., 0.)
MBsst <- rxtracto_3D(dataInfo, parameter, xcoord = xcoord, ycoord = ycoord,
                       tcoord = tcoord, zcoord = zcoord)

# low resolution selected to keep time to render down
p <- plotBBox(MBsst, plotColor = 'temperature', maxpixels = 300)

#> Loading required package: maps

#> grid object contains more than 300 pixels

#> increase `maxpixels` for a finer resolution